splunk/requirements.txt
httplib2
configparser
pathlib
mysql-connector
